storybook-agent/pyproject.toml
[project]
name = "storybook-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.0.0",
    "google-genai>=1.31.0",
    "python-dotenv>=1.1.1",
    "jupyter>=1.1.1",
    "ipykernel>=6.29.0",
]

[tool.hatch.build.targets.wheel]
packages = ["storybook"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
